14. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -13,15 +13,14 @@
 int main(void)
 {
 	int n = 50;
-	int i = 1;
-	int o = 2;
+	long i = 1;
+	long o = 2;
 	int z = 2;
 
 	printf("%d, %d", i, o);
-	
 	while (z < n)
 	{
-		int s = i + o;
+		long s = i + o;
 	       printf(", %d", s);
 	       i = o;
 		o = s;
